新增考试功能，AI接入

+++ back-end-spring-cloud/general-service/src/main/java/org/example/generalservice/controller/ScoreRecordController.java
@@ -91,7 +91,7 @@ public class ScoreRecordController {
      * 删除成绩记录
      */
     @DeleteMapping("/delete/{id}")
-    public RestBean<String> deleteScore(@PathVariable Integer id) {
+    public RestBean<String> deleteScore(@PathVariable Long id) {
         log.info("删除成绩记录: id={}", id);
         if (scoreRecordService.removeById(id)) {
             return RestBean.success("删除成功");
